FrameStatsCollector
