+++ core/frontend/src/Viewport.ts
@@ -38,7 +38,8 @@ import { ToolTipOptions } from "./NotificationManager";
 import { PerModelCategoryVisibility } from "./PerModelCategoryVisibility";
 import { Decorations } from "./render/Decorations";
 import { FeatureSymbology } from "./render/FeatureSymbology";
-import { FrameStats, FrameStatsCollector } from "./render/FrameStats";
+import { FrameStats } from "./render/FrameStats";
+import { FrameStatsCollector } from "./internal/render/FrameStatsCollector";
 import { AnimationBranchStates } from "./render/GraphicBranch";
 import { Pixel } from "./render/Pixel";
 import { GraphicList } from "./render/RenderGraphic";
